// client/src/constants/actionTypes.ts
/* eslint-disable no-unused-vars */
export enum PetActionTypes {
  FETCH_PETS_REQUEST = "FETCH_PETS_REQUEST",
  FETCH_PETS_SUCCESS = "FETCH_PETS_SUCCESS",
  FETCH_PETS_FAILURE = "FETCH_PETS_FAILURE",

  REMOVE_PET_REQUEST = "REMOVE_PET_REQUEST",
  REMOVE_PET_SUCCESS = "REMOVE_PET_SUCCESS",
  REMOVE_PET_FAILURE = "REMOVE_PET_FAILURE",
  REMOVE_PET_FROM_USER_PROFILE = "REMOVE_PET_FORM_USER_PROFILE",

  CREATE_PET_REQUEST = "CREATE_PET_REQUEST",
  CREATE_PET_SUCCESS = "CREATE_PET_SUCCESS",
  CREATE_PET_FAILURE = "CREATE_PET_FAILURE",

  UPDATE_PET_REQUEST = "UPDATE_PET_REQUEST",
  UPDATE_PET_SUCCESS = "UPDATE_PET_SUCCESS",
  UPDATE_PET_FAILURE = "UPDATE_PET_FAILURE",

  UPDATE_PET_IN_LIST = "UPDATE_PET_IN_LIST",
  UPDATE_PET_IN_USER_PROFILE_LIST = "UPDATE_PET_IN_USER_PROFILE_LIST",

  ADD_TO_FAVORITES = "ADD_TO_FAVORITES",
  REMOVE_FROM_FAVORITES = "REMOVE_FROM_FAVORITES",

  SEARCH_PETS = "SEARCH_PETS",
  FILTER_PETS = "FILTER_PETS",
}

export enum CartActionTypes {
  ADD_TO_CART = "ADD_TO_CART",
  REMOVE_FROM_CART = "REMOVE_FROM_CART",
  CLEAR_CART = "CLEAR_CART",
}

export enum AuthActionTypes {
  REGISTER_REQUEST = "REGISTER_REQUEST",
  REGISTER_SUCCESS = "REGISTER_SUCCESS",
  REGISTER_FAILURE = "REGISTER_FAILURE",
  LOGIN_REQUEST = "LOGIN_REQUEST",
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_FAILURE = "LOGIN_FAILURE",
  LOGOUT = "LOGOUT",
  LOAD_USER_FAILURE = "LOAD_USER_FAILURE",
  LOAD_USER_REQUEST = "LOAD_USER_REQUEST",
  LOAD_USER_SUCCESS = "LOAD_USER_SUCCESS",
}

export enum UserActionTypes {
  GET_USER_PROFILE_INFO_REQUEST = "GET_USER_PROFILE_INFO_REQUEST",
  GET_USER_PROFILE_INFO_SUCCESS = "GET_USER_PROFILE_INFO_SUCCESS",
  GET_USER_PROFILE_INFO_FAILURE = "GET_USER_PROFILE_INFO_FAILURE",
}